backend/requirements.txt
django>=5.1,<5.2
python-decouple>=3.7,<=3.8
djangorestframework>=3.12,<3.16
psycopg2-binary>=2.9,<3.0
django-cors-headers>=3.0,<4.0
django-filter>=24.0,<=24.3
pytest-django>=4.9,<5.0